% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_bulk.R
\name{check_interactive}
\alias{check_interactive}
\title{Check if R is running interactively}
\usage{
check_interactive()
}
\value{
Logical indicating whether R is running interactively
}
\description{
Wrapper around base::interactive() to make the function testable.
This function exists primarily to facilitate testing of interactive features.
}
\keyword{internal}
