% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_bulk.R
\name{download_file}
\alias{download_file}
\title{Download a file using utils::download.file}
\usage{
download_file(url, destfile, quiet)
}
\arguments{
\item{url}{URL of file to download}

\item{destfile}{Destination file path}

\item{quiet}{Whether to suppress messages}
}
\value{
Invisibly returns the status code from download.file
}
\description{
Wrapper around utils::download.file to facilitate testing.
}
\keyword{internal}
