% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_bulk.R
\name{get_response_headers}
\alias{get_response_headers}
\title{Get response headers from a URL}
\usage{
get_response_headers(file_url)
}
\arguments{
\item{file_url}{URL to request headers from}
}
\value{
List of response headers
}
\description{
Get response headers from a URL
}
