% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_bulk.R
\name{process_bulk_data}
\alias{process_bulk_data}
\title{Process bulk data into tidy format}
\usage{
process_bulk_data(bulk_raw)
}
\arguments{
\item{bulk_raw}{Raw data frame from Excel file}
}
\value{
Processed tibble in tidy format
}
\description{
Process bulk data into tidy format
}
