% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_bulk.R
\name{prompt_user}
\alias{prompt_user}
\title{Prompt a user with a question}
\usage{
prompt_user(prompt)
}
\description{
Wrapper around base::readline to facilitate testing. Cannot be tested
because of the base binding.
}
\keyword{internal}
