% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_bulk.R
\name{read_bulk_file}
\alias{read_bulk_file}
\title{Read bulk file and determine column types}
\usage{
read_bulk_file(file_path)
}
\arguments{
\item{file_path}{Path to Excel file}
}
\value{
Raw data frame from Excel file
}
\description{
Read bulk file and determine column types
}
