% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids_bulk.R
\name{validate_file}
\alias{validate_file}
\title{Validate that downloaded file exists and is not empty}
\usage{
validate_file(file_path)
}
\arguments{
\item{file_path}{Path to file to validate}
}
\description{
Validate that downloaded file exists and is not empty
}
