% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbids-package.R
\docType{package}
\name{wbids-package}
\alias{wbids}
\alias{wbids-package}
\title{wbids: Work seamlessly with World Bank International Debt Statistics (IDS)}
\description{
A modern, flexible interface for accessing the World Bank’s International Debt Statistics (IDS) \url{https://datacatalog.worldbank.org/search/dataset/0038015}. 'wbids' allows users to download, process, and analyze IDS series for multiple geographies, counterparts and specific time periods. The package is designed to work seamlessly with World Development Indicators (WDI) provided through the 'wbwdi' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://teal-insights.github.io/r-wbids}
  \item \url{https://github.com/teal-insights/r-wbids}
  \item Report bugs at \url{https://github.com/teal-insights/r-wbids/issues}
}

}
\author{
\strong{Maintainer}: Teal Emery \email{lte@tealinsights.com}

Authors:
\itemize{
  \item Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID})
}

Other contributors:
\itemize{
  \item Teal Insights [copyright holder]
}

}
\keyword{internal}
