% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbids-package.R
\docType{package}
\name{wbids-package}
\alias{wbids}
\alias{wbids-package}
\title{wbids: Work seamlessly with World Bank International Debt Statistics (IDS)}
\description{
Work seamlessly with World Bank International Debt Statistics (IDS). A a link to the World Bank IDS API docs should be here. This package is a product of Teal-Insight and is not sponsored by or affiliated with the World Bank in any way, except for the use of the World Bank International Debt Statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://teal-insights.github.io/r-wbids}
  \item \url{https://github.com/Teal-Insights/r-wbids}
  \item Report bugs at \url{https://github.com/Teal-Insights/r-wbids/issues}
}

}
\author{
\strong{Maintainer}: Teal Emery \email{lte@tealinsights.com}

Other contributors:
\itemize{
  \item Teal Insights [copyright holder]
}

}
\keyword{internal}
